backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
pymongo==4.3.3
python-dotenv==1.0.0
pyjwt==2.6.0
werkzeug==2.2.3
razorpay==1.3.0
python-decouple==3.8
bcrypt==4.0.1